+++ src/Clywell.Core.Security/PermissionRequirement.cs
@@ -1,6 +1,0 @@
-namespace Clywell.Core.Security;
-
-public sealed class PermissionRequirement(string permission) : IAuthorizationRequirement
-{
-    public string Permission { get; } = permission;
-}